project/requirements.txt
django==3.2.20
django_select2
django-allauth
django-crispy-forms
django-widget-tweaks
pandas==0.24.2
reportlab==3.6.6
xhtml2pdf
psycopg2
python-dotenv
django-import_export
django_cpf_cnpj
num2words
xmltodict
pynfe